Refactor pagination to behavior and apply to all work types
