+++ app/controllers/hyrax/videos_controller.rb
@@ -10,6 +10,7 @@ module Hyrax
     include Hyrax::BreadcrumbsForWorks
     include OregonDigital::DownloadControllerBehavior
     prepend OregonDigital::WorksControllerBehavior
+    include OregonDigital::WorkRelationPaginationBehavior
     self.curation_concern_type = ::Video
 
     # Override the way Hyrax's works present iiif manifests
